+++ js/background.js
@@ -18,11 +18,7 @@
 
 // keyboard shortcuts
 chrome.commands.onCommand.addListener(function(command) {
-	if (command === 'savePage' || command === 'updatePage') {
-		chrome.tabs.query({active: true, currentWindow: true}, function(tabs) {
-			chrome.tabs.sendMessage(tabs[0].id, {type: command}, handleResponse);
-		});
-	} else if (command === 'assignToMe') {
+	if (command === 'assignToMe') {
 		chrome.storage.sync.get(['debug', 'userId', 'fullName', 'groupId', 'groupName'], function(items) {
 			if (chrome.runtime.lastError) {
 				console.warn('SNAFU Sync Get Error: %s', chrome.runtime.lastError.message);
@@ -48,6 +44,10 @@ chrome.commands.onCommand.addListener(function(command) {
 					chrome.tabs.sendMessage(tabs[0].id, messageData, handleResponse);
 				});
 			}
+		});
+	} else {
+		chrome.tabs.query({active: true, currentWindow: true}, function(tabs) {
+			chrome.tabs.sendMessage(tabs[0].id, {type: command}, handleResponse);
 		});
 	}
 });