+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### ToDo
+- Detect Equipment Move/Remove tasks based on their auto-generated short description.
+
 ## 1.1.0 (2017-06-02)
 ### Added
 - Scheduled option to put a ticket on hold with an appointment message.